More name changes
In order to change things about a model that has a scaffold, if you want to modify the model, you must also modify the scaffold. You can do this by manually doing this, or by deleting the old scaffold and creating a new one. Remember that you have to make a migration as well when you do this.